Define Behavior Specification (#3)
We create the first set of scenarios in the `define` feature file for Graql. This is the high level specification of the behaviors of Graql, from the user's point of view.

This PR creates new file: `define.feature`, which is a mostly fleshed out spec (some scenarios are left empty for now, are mostly duplicates of very similar scenarios) of how a `define` statement should operate.