Refactor Pipe to Triple Quote characters for Graql Queries (#19)
Currently, we use | |  for multiline queries, which are joined together in the corresponding Graql step BDD implementation. The issue with this is that the UX from the user side seems like multiple queries can actually be written over multiple lines as well.

However, docstrings can be used for this particular data input as well:
"""
asdfasf
"""

Note that the triple quotes must be on their own lines.
This format ensures that only 1 query can be written in all the places queries are currently used as step arguments.