Replaced the term 'DataType' with 'ValueType' (#35)
## What is the goal of this PR?

We are replacing the term `DataType` in `AttributeType` with `ValueType`). Previously, an `AttributeType` has a `DataType`. And an `Attribute` has `Value`. The type of `Value` is defined by `DataType`. We can see the inconsistency there. Now, `AttributeType` has `ValueType`, and `Attribute` has `Value`. The type of `Value` is `ValueType`. So it's all consistent.

## What are the changes implemented in this PR?

- Renamed the `DataType` class to `ValueType`
- Renamed method names
- Renamed method arguments
- Renamed variable names
- Updated comments

Additionally, we also cleaned up dependencies.
1) We declare direct dependencies onto `@graknlabs_graql`, `@graknlabs_client_java`, and `@graknlabs_grakn_core`
2) We only run tests in CI against `grakn-core-server` rather than `grakn-core-all`, as we do not need the full distribution with Grakn Console. This also allowed us to trim down the dependencies declared in `WORKSPACE`.