Migrate ReasoningIT to BDD (#81)
## What is the goal of this PR?

To identify and de-duplicate integration tests that test the fundamentals of Reasoner, categorise them appropriately and migrate them to BDD scenarios.

This PR focuses on ReasoningIT in grakn.

This PR forms part of the work towards https://github.com/graknlabs/grakn/issues/5721

## What are the changes implemented in this PR?

Added tests that were migrated from the relevant IT files in grakn. Also moved some tests around, so that everything is in the right place.