Increased test coverage for rule validation (#105)
## What is the goal of this PR?

Making sure all rule validation mechanism are covered by the BDD tests. This has been kept in rule-validation.feature. The aim is to have a thorough testing to give us confidence when we switch the reasoner engine in 2.0.

## What are the changes implemented in this PR?

Added more tests in rule-validation.feature file, and also tidied the tests up a little by added nickname into the default define block in order to reduce repetition.