Connection scenario corrections (#124)
## What is the goal of this PR?

We updated the scenario "write data in schema session" to properly reflect intended behaviour- it should throw when committed, not on the write itself. 

## What are the changes implemented in this PR?

Wording of the graql define step in database.feature has been corrected to the new format
An additional commit step has been added to the write data in schema session scenario, with the throw moved to that step.