Simple define feature fixes (#146)
## What is the goal of this PR?

We had a number of behavioural tests set as ignored in the define feature as a result of them failing. A large number of these issues were a consequence of typograpical errors or . We changed these tests to be correct, and they now pass.

## What are the changes implemented in this PR?

We unignored several failing tests in `define.feature`, fixing their labels where appropriate and introducing the unimplemented `an existing relation type cannot be converted to abstract if it has existing instances` scenario that was present in title only.