Test non-deterministic results on negated relation which is both inferred and retrieved (#220)
## What is the goal of this PR?
A rule containing a negated relation was reported to (unsoundly) return 2 or 3 answers non-deterministically. (see [typedb issue#6500](https://github.com/vaticle/typedb/issues/6500)). The relation could be retrieved from the database or inferred using rules. This behaviour isn't shown by release 2.10.0 but can be reproduced on earlier versions.

## What are the changes implemented in this PR?
To prevent regression, the scenario in the issue is converted to a behaviour test and added to the negation test-suite.