Add scenario ensuring buffered iids in overrides are translated (#280)
## Usage and product changes
Adds test for a bug which creates an invalid override when a committed
role-type is overridden with a newly defined role-type.

## Implementation
Adds test for a bug which creates an invalid override when a committed
role-type is overridden with a newly defined role-type. The bug was
caused by Persisted TypeEdges assuming any overridden IID would also be
persisted.