Concept Type BDDs for all the types, edges, its annotations and ordering (#287)
## Usage and product changes
We refactor already existing BDDs for `concept/type` package, adding new
base cases of type manipulations and refactoring already existing cases
for better scalability in the future.

We add tests for all the annotations from `3.0` and basic lists (of
roles and of owns) manipulations for:
* entity type;
* relation type;
* attribute type + relates (roles);
* owns;
* plays.

The ultimate goal of this PR is to have decent coverage for schema and
refactor the `concept/type` tests so this format could be reused in
`concept/thing` later on, additionally being available for small
additions on the schema level.

Moreover, all the files now tend to have structure for easier test cases
search:
* common steps
* common lists steps (if applicable)
* annotations common
* special cases for specific annotations
* annotations combinations
* -- you can insert your new feature here --

## Implementation
I won't put many additional comments inside the steps declarations, here
is the whole process of additions and refactoring and my logic behind
it:

**Setup:**
* I created new files: `owns.feature` and `plays.feature`.
* I moved all `entitytype` and `relationtype` tests that check `owns`
behavior to `owns.feature` and tests that check `plays` behavior to
`plays.feature`.
* `relates` behavior (`roles` on the concept level) remains a part of
the `attributetype` file.

**Owns:**
* I added base cases of the possibility of owning something from a root
type for all the possible combinations.
* I refactored the moved from `entitytype` and `relationtype` files
tests, putting most of them under a `Scenario Outline` with `Examples`
of entities and relations (I add these entities and relations to the
`Background` step so they always exist for Examples substitutions). Some
of them are left separated so as not to spam `entitytype`/`relationtype`
X `attribute-valuetype` permutations. I also added some more declaration
basic and inheritance cases.
* I duplicated everything mentioned above for lists: `owns smth[]` as
it's easier to add special cases for lists later, and `Examples` for
`Scenario Outlines` would cause too much trouble for this case.
* I covered all the possible `annotation`s: base cases of declaration
(with fails), inheritance, redeclarations, etc.
* I covered special cases for specific annotations with all the possible
attribute value types (that's why they are separate).
Additionally, I added negative cases for unsupported annotations and
positive-negative cases with annotation combinations!

**Plays:**
I copied the base of `owns` tests and followed the same ideas. The main
differences here:
* I only added some tests for playing a role from a list of roles (`me
sub relation, relates smth[]; ... plays me:smth`), no tests for lists of
`plays` as we don't support it _yet_.
* The only compatible annotation for plays is `@card`, so there are much
fewer cases for annotations
* I spent less time on refactoring existing tests from old files

**Entity type and Relation type**:
Followed a similar approach as for `plays`, haven't refactored old tests
much, but tried to cover annotations decently.
Relation types also received additional special tests for `relates`
(`roles`) and lists of `relates`.

**Attribute type**:
Closer to `owns` for easier changes and scalability regarding
`annotation`s, but active in using the old tests for simple cases.

**Addons**:
* I've tried to put steps declarations to a similar form of `type(name)
get smth(name) get smthelse contains:`, etc. Now for the schema level,
but I'll be able to propagate this refactoring everywhere once we
approve it.
* There are many `TODO`s that require our revisiting later / while
fixing the code based on these tests / right now as a part of the
review. I'll highlight the main ones by my comments so we can address
them here. All the other `TODO`s are much easier to be addressed while
working with code.