Schema commit time validations (#290)
## Usage and product changes
We fix old `concept/schema-validation` behaviour tests, reformatting
them to the new `ConceptAPI` steps declarations and changing small
details in its logic to follow the new validation rules of `3.0`.

We also fix a number of new `concept/type` tests that expected
`transaction commits; fails;` results based on the final versions of
commit time validations of the `3.0` server. It is expected that every
commit-time validation is now covered by BDD tests.

Additionally, this PR presents new steps for `@values`, `@ranges`,
`@card` annotations (with a separate check of `get cardinality`
constraint), and other small things that were forgotten in the past.

## Implementation
I'd like to move `schema-validation` tests to different `type` files as
they have just the same meaning and it's hard to split operation time
validations with commit time validations (schema validations in
migration). Right now all the old tests are still in the old file for
easier review.

For more details, check out the comments in the changed files.