Rewrite and add 3.0 tests for concept/thing and concept/migration (ex. validation) (#292)
## Usage and product changes
We rename `validation` to `migration` to collect all the
`migration`-related tests in this directory.
`data-validation` has more data validation-related tests for schema
migration, and `migration` is a set of tests to check the results of
schema migration, resulting cleanups and states.
`rule` is now `function` and these tests should be refactored and
enhanced when we fully introduce functions (it will probably require
TypeQL, and we'll place it in `concept/migration`... which is a bit
strange, but should be fine).

We move all the data-related tests from `type` to `migration`. We
rewrite all the existing tests from `migration` and `thing`, adding new
cases with new annotations and features (like ordering). It is expected
that all the new features of 3.0 for concept api are now extensively
covered by BDD tests.

Misc:
* removed root types;
* changed `@card` format from `@card(1, 2)` to `@card(1..2)`