Introduce 3.0 fetch BDD (#312)
## Usage and product changes
We introduce new BDD tests for fetch. These tests cover all forms of
`fetch` queries from the 3.0 spec (excluding lists of attributes).

Some of the cases are commented out with explicit `TODO`s, as there are
expressions or functions that are being fixed in other branches. This
will require revisiting this file later to uncomment these cases.