Add a test for driver with database background creation (#314)
## Usage and product changes
We encountered a bug when the Rust driver crashed on transaction opening
for a database created by another driver in background.
To test such cases with multiple drivers working in parallel, we add a
BDD scenario describing this situation.

## Implementation
We don't want to overcomplicate the implementation of the drivers BDDs
by managing multiple parallel driver connections as we don't have many
things to test in this setup at the moment. That's why we introduce a
new `in background,` category of steps, which means "create a new
temporary driver and do something without affecting the main
connection".