Introduce tests for update queries (#338)
## Usage and product changes
Introduce tests for update queries. These queries are used to make sure
that a type capability has a specific value: either an attribute or a
role player.
 
The old tests in the `update.feature` using `delete + insert` were
either moved into the `pipelines` or updated to contain two compared
versions: `delete + insert` and `update`. In the meantime, the test file
is now full of new tests for the new feature:
1. Validations of incorrect usages of the queries
2. Updating `has`: introduction of a `has` without having an old
capability, updates of existing capabilities, validation for a
multi-layered attribute `sub` hierarchy, cardinality validations, and
more.
3. Updating `links`: same as in the previous point, but with specifics
of roleplayers.
4. Combinations and edge cases.

## Implementation