Add a uniqueness validation test for disconnected owners (#356)
## Usage and product changes
We add an additional test for operation time `put has` uniqueness
validation, where:
* Three entity types' hierarchies are created, each with the same `owns
attr @unique` ownership. These hierarchies should not affect each other
* These types are created in a messy order, so their children are placed
in a mixed order in the storage. This allows us to check that the actual
implementation of this validation does not confuse types (while this
test is added after a very specific bug mentioned below, it is still a
useful thing we could use more... Sometimes)

Details:
https://github.com/typedb/typedb/pull/7409#discussion_r2019001159

## Implementation